refactor: Moved `_get_storage_client` and `get_bytes` to utilities
- Resolves Circular dependency
