+++ google/cloud/documentai_toolbox/converters/config/converter_helpers.py
@@ -28,7 +28,9 @@ from google.cloud.documentai_toolbox.converters.config.blocks import (
     _load_blocks_from_schema,
 )
 
-from google.cloud.documentai_toolbox import document, constants
+from google.cloud.documentai_toolbox import constants
+from google.cloud.documentai_toolbox.utilities import utilities
+
 from google.cloud import documentai, storage
 
 
@@ -86,7 +88,6 @@ def _get_entity_content(
     entity_id = 0
 
     for block in blocks:
-
         docai_entity = documentai.Document.Entity()
         if block.confidence:
             docai_entity.confidence = block.confidence
@@ -233,7 +234,7 @@ def _get_bytes(
 
     """
 
-    storage_client = document._get_storage_client()
+    storage_client = utilities._get_storage_client()
     bucket = storage_client.bucket(bucket_name=bucket_name)
     blobs = storage_client.list_blobs(bucket_or_name=bucket_name, prefix=prefix)
 
@@ -287,7 +288,7 @@ def _upload_file(
         None.
 
     """
-    storage_client = document._get_storage_client()
+    storage_client = utilities._get_storage_client()
     bucket = storage_client.bucket(bucket_name)
     blob = bucket.blob(output_prefix)
 
@@ -494,7 +495,7 @@ def _convert_documents_with_config(
     if file_check:
         raise ValueError("gcs_prefix cannot contain file types")
 
-    storage_client = document._get_storage_client()
+    storage_client = utilities._get_storage_client()
 
     blob_list = storage_client.list_blobs(input_bucket, prefix=input_prefix)
